requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
